requirements.txt
seleniumbase
python-dotenv 